Move RatioOps and NumberOps to numerics.  Move numerics algebras to numerics module.
