+++ blockchain/src/main/scala/co/topl/blockchain/PrivateTestnet.scala
@@ -5,6 +5,7 @@ import co.topl.crypto.hash.Blake2b256
 import co.topl.models._
 import co.topl.models.utility.HasLength.instances._
 import co.topl.models.utility._
+import co.topl.numerics.implicits._
 import co.topl.typeclasses.implicits._
 
 object PrivateTestnet {
