+++ consensus/src/main/scala/co/topl/consensus/interpreters/LeaderElectionValidation.scala
@@ -2,14 +2,15 @@ package co.topl.consensus.interpreters
 
 import cats.effect.Sync
 import cats.implicits._
-import co.topl.algebras.{Exp, Log1p, UnsafeResource}
+import co.topl.algebras.UnsafeResource
 import co.topl.codecs.bytes.typeclasses.Signable
 import co.topl.consensus.algebras.LeaderElectionValidationAlgebra
 import co.topl.crypto.signing.Ed25519VRF
 import co.topl.crypto.hash.Blake2b512
 import co.topl.models._
 import co.topl.models.utility.Ratio
-import co.topl.typeclasses.implicits._
+import co.topl.numerics.algebras.{Exp, Log1p}
+import co.topl.numerics.implicits._
 import scalacache.caffeine.CaffeineCache
 
 object LeaderElectionValidation {
