+++ consensus/src/test/scala/co/topl/consensus/interpreters/ConsensusDataEventSourcedStateSpec.scala
@@ -10,6 +10,7 @@ import co.topl.codecs.bytes.typeclasses.implicits._
 import co.topl.eventtree.ParentChildTree
 import co.topl.models.ModelGenerators._
 import co.topl.models._
+import co.topl.numerics.implicits._
 import co.topl.typeclasses.implicits._
 import munit.{CatsEffectSuite, ScalaCheckEffectSuite}
 import org.scalamock.munit.AsyncMockFactory
