+++ networking/src/main/scala/co/topl/networking/blockchain/BlockchainPeerClient.scala
@@ -9,6 +9,7 @@ import cats.implicits._
 import co.topl.models._
 import co.topl.models.utility.Ratio
 import co.topl.networking.p2p.ConnectedPeer
+import co.topl.numerics.implicits._
 import co.topl.typeclasses.implicits._
 import org.typelevel.log4cats.Logger
