+++ numerics/src/main/scala/co/topl/numerics/package.scala
@@ -1,0 +1,5 @@
+package co.topl
+
+package object numerics {
+  object implicits extends NumberOps with RatioOps.Implicits
+}
