+++ typeclasses/src/main/scala/co/topl/typeclasses/package.scala
@@ -9,7 +9,6 @@ package object typeclasses {
       with ContainsTransactions.ToContainsTransactionsOps
       with ContainsTransactionIds.Instances
       with ContainsTransactionIds.ToContainsTransactionIdsOps
-      with RatioOps.Implicits
       with ShowInstances
       with EqInstances
       with SpendingAddressable.Instances
@@ -18,5 +17,4 @@ package object typeclasses {
       with NonEmpty.Instances
       with TransactionOps.Instances
       with IdentityOps
-      with NumberOps
 }
